requirements_dev.txt
alabaster==0.7.12
appdirs==1.4.4
argh==0.26.2
Babel==2.8.0
bleach==3.2.1
bump2version==0.5.11
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
Click==7.0
colorama==0.4.3
coverage==4.5.4
cryptography==3.1.1
digi-xbee==1.3.0
distlib==0.3.1
docutils==0.16
entrypoints==0.3
filelock==3.0.12
flake8==3.7.8
idna==2.10
imagesize==1.2.0
importlib-metadata==0.23
jeepney==0.4.3
Jinja2==2.11.2
keyring==21.4.0
LowCostSmartFarmHub
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.4
paho-mqtt==1.5.1
pathtools==0.1.2
pkg-resources==0.0.0
pkginfo==1.5.0.1
pluggy==0.13.1
py==1.9.0
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
Pygments==2.7.1
pyparsing==2.4.7
pyserial==3.4
pytz==2020.1
PyYAML==5.3.1
readme-renderer==26.0
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
RPi.version===2.0
SecretStorage==3.1.2
six==1.15.0
snowballstemmer==2.0.0
Sphinx==1.8.5
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.4
sphinxcontrib-napoleon===0.7
srp==1.0.16
toml==0.10.1
tox==3.14.0
tqdm==4.50.0
twine==3.2.0
urllib3==1.25.10
virtualenv==20.0.32
watchdog==0.9.0
webencodings==0.5.1
zipp==3.2.0
